+++ Drawingtoon/Drawingtoon/Shared/Service/Network/HTTPMethod.swift
@@ -1,0 +1,12 @@
+//
+//  HTTPMethod.swift
+//  Drawingtoon
+//
+//  Created by Heejung Yang on 9/25/25.
+//
+
+import Foundation
+
+public enum HTTPMethod: String {
+    case GET, POST, PUT, PATCH, DELETE
+}
